Moving some types around - make sure app types are all correctly filed away.
