+++ packages/server/src/api/controllers/templates.ts
@@ -1,13 +1,17 @@
 import nodeFetch from "node-fetch"
 import { downloadTemplate as dlTemplate } from "../../utilities/fileSystem"
 import env from "../../environment"
-import { BBContext } from "@budibase/types"
+import {
+  DownloadTemplateResponse,
+  FetchTemplateResponse,
+  UserCtx,
+} from "@budibase/types"
 
 // development flag, can be used to test against templates exported locally
 const DEFAULT_TEMPLATES_BUCKET =
   "prod-budi-templates.s3-eu-west-1.amazonaws.com"
 
-export async function fetch(ctx: BBContext) {
+export async function fetch(ctx: UserCtx<void, FetchTemplateResponse>) {
   let type = env.TEMPLATE_REPOSITORY
   let response,
     error = false
@@ -32,7 +36,9 @@ export async function fetch(ctx: BBContext) {
 
 // can't currently test this, have to ignore from coverage
 /* istanbul ignore next */
-export async function downloadTemplate(ctx: BBContext) {
+export async function downloadTemplate(
+  ctx: UserCtx<void, DownloadTemplateResponse>
+) {
   const { type, name } = ctx.params
 
   await dlTemplate(type, name)
