+++ packages/types/src/api/web/app/application.ts
@@ -1,5 +1,5 @@
-import type { PlanType } from "../../sdk"
-import type { Layout, App, Screen } from "../../documents"
+import type { PlanType } from "../../../sdk"
+import type { Layout, App, Screen } from "../../../documents"
 import { ReadStream } from "fs"
 
 export interface SyncAppResponse {
