+++ packages/types/src/api/web/app/index.ts
@@ -13,3 +13,7 @@ export * from "./integration"
 export * from "./metadata"
 export * from "./query"
 export * from "./screen"
+export * from "./application"
+export * from "./layout"
+export * from "./deployment"
+export * from "./role"
