+++ packages/types/src/api/web/index.ts
@@ -1,4 +1,3 @@
-export * from "./application"
 export * from "./analytics"
 export * from "./auth"
 export * from "./user"
@@ -11,10 +10,7 @@ export * from "./global"
 export * from "./pagination"
 export * from "./searchFilter"
 export * from "./cookies"
-export * from "./automation"
-export * from "./layout"
-export * from "./role"
 export * from "./plugins"
 export * from "./apikeys"
-export * from "./deployment"
 export * from "./dev"
+export * from "./template"
