+++ packages/types/src/api/web/template.ts
@@ -1,0 +1,5 @@
+export interface FetchTemplateResponse {}
+
+export interface DownloadTemplateResponse {
+  message: string
+}
